# Conflicts:
#	conf/deployment_configs.json
#	package-lock.json
#	package.json
#	src/app.js
#	src/bootup/init/init_mysql.js
#	src/bootup/init/setup_graceful_close.js
#	src/bootup/routes/load_logic_routes.js
#	src/controllers/files.js
#	src/controllers/posts.js
#	src/controllers/project_descriptors_recommendation.js
#	src/controllers/records.js
#	src/controllers/timeline.js
#	src/models/directory_structure/information_element.js
#	src/models/recommendation/interaction.js
#	src/utils/logger.js
#	test/index.Test.dev.js
#	test/index.Test.js
#	test/mockdata/projects/public_project_for_demouser2.js
#	test/routes/posts/posts/routes.posts.posts.Test.js
#	test/units/projects/createProjects.Unit.js
#	test/units/social/createSocialDendroTimelineWithPostsAndShares.Unit.js
#	test/units/social/shareSomePosts.Unit.js

+++ conf/scripts/install.sh
@@ -3,12 +3,20 @@
 INITIAL_DIR=`pwd`
 NODE_VERSION=`cat .nvmrc`
 
-# install text extraction dependencies
-if [ "$(uname)" == "Darwin" ]; then
-    brew cask install xquartz && brew install ghostscript xpdf tesseract imagemagick && brew cask install pdftotext
-elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
-    echo "I need your sudo password for installing text extraction dependencies..."
-    sudo apt-get -y -f install poppler-utils antiword unrtf tesseract-ocr
+# check to see if text extraction tools need to be installed
+if [[ tesseract > /dev/null ]]
+then
+    # install text extraction dependencies
+    if [ "$(uname)" == "Darwin" ]; then
+        brew cask install xquartz && brew install ghostscript xpdf tesseract imagemagick@6 && brew cask install pdftotext
+        brew tap caskroom/versions
+        brew cask install java8
+    elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
+        if [[ $(dpkg -l "poppler-utils" "antiword" "unrtf" "tesseract-ocr") ]]; then
+            echo "I need your sudo password for installing text extraction dependencies..."
+            sudo apt-get -y -f install poppler-utils antiword unrtf tesseract-ocr
+        fi
+    fi
 fi
 
 echo "Installing Dendro in $INITIAL_DIR with username $(whoami) and Node $NODE_VERSION"
@@ -21,24 +29,27 @@ else
     chown -R "$(whoami)" "$HOME/.nvm"
 
     #install NVM, Node, Node Automatic Version switcher
-    curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.1/install.sh | bash &&
+    curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.8/install.sh | bash &&
     export NVM_DIR="$HOME/.nvm" &&
     [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh" && # This loads nvm
 
     source $HOME/.bash_profile
-    nvm install "$NODE_VERSION"
-    nvm use "$NODE_VERSION"
+#    nvm install "$NODE_VERSION"
+#    nvm use --delete-prefix "$NODE_VERSION" --silent
 
     echo "Installing Node Version $NODE_VERSION during install script!!"
     nvm install $NODE_VERSION &&
-    nvm use $NODE_VERSION &&
+    nvm use --delete-prefix $NODE_VERSION --silent &&
     echo "loaded NVM."
-
-    #update npm
-    npm install npm@latest -g
 
     #clear npm cache
     npm cache clean --force
+
+    #update npm (force 5.6.0 because of write after end issue: https://github.com/npm/npm/issues/19989)
+    npm i -g npm@5.6.0
+
+    #install nyc
+    npm i -g nyc
 
     #delete node_modules folder
     rm -rf node_modules
@@ -47,7 +58,7 @@ else
     chown -R "$(whoami)" "$HOME/.nvm"
 
     #install preliminary dependencies
-    npm i -g npm && npm i -g grunt && npm install gulp-cli -g && npm install bower -g && npm install pm2 -g && npm install -g npm-check-updates
+    npm i -g grunt && npm install gulp-cli -g && npm install bower -g && npm install pm2 -g && npm install -g npm-check-updates
 
     #install dependencies. Will also run bower install whenever needed
     npm install #this is needed when running npm install with sudo to install global modules
